- GPUTexture 접근비용을 줄임
