+++ src/material/RedStandardMaterial.js
@@ -2,7 +2,7 @@
  *   RedGPU - MIT License
  *   Copyright (c) 2019 ~ By RedCamel( webseon@gmail.com )
  *   issue : https://github.com/redcamel/RedGPU/issues
- *   Last modification time of this file - 2019.12.14 17:33:43
+ *   Last modification time of this file - 2019.12.17 9:45:10
  *
  */
 
@@ -166,7 +166,7 @@ export default class RedStandardMaterial extends RedMix.mix(
 
 	checkTexture(texture, textureName) {
 		if (texture) {
-			if (texture.GPUTexture) {
+			if (texture._GPUTexture) {
 				switch (textureName) {
 					case 'diffuseTexture' :
 						this._diffuseTexture = texture;
@@ -184,7 +184,7 @@ export default class RedStandardMaterial extends RedMix.mix(
 						this._displacementTexture = texture;
 						break
 				}
-				if (RedGPUContext.useDebugConsole) 	console.log("로딩완료or로딩에러확인 textureName", textureName, texture ? texture.GPUTexture : '');
+				if (RedGPUContext.useDebugConsole) 	console.log("로딩완료or로딩에러확인 textureName", textureName, texture ? texture._GPUTexture : '');
 
 				cancelAnimationFrame(this.#raf)
 				this.#raf = requestAnimationFrame(_=>{this.needResetBindingInfo = true})
@@ -212,7 +212,7 @@ export default class RedStandardMaterial extends RedMix.mix(
 			{binding: 1, resource: this.sampler.GPUSampler},
 			{
 				binding: 2,
-				resource: this._displacementTexture ? this._displacementTexture.GPUTextureView : this.redGPUContext.state.emptyTextureView
+				resource: this._displacementTexture ? this._displacementTexture._GPUTextureView : this.redGPUContext.state.emptyTextureView
 			},
 			{
 				binding: 3,
@@ -225,19 +225,19 @@ export default class RedStandardMaterial extends RedMix.mix(
 			{binding: 4, resource: this.sampler.GPUSampler},
 			{
 				binding: 5,
-				resource: this._diffuseTexture ? this._diffuseTexture.GPUTextureView : this.redGPUContext.state.emptyTextureView
+				resource: this._diffuseTexture ? this._diffuseTexture._GPUTextureView : this.redGPUContext.state.emptyTextureView
 			},
 			{
 				binding: 6,
-				resource: this._normalTexture ? this._normalTexture.GPUTextureView : this.redGPUContext.state.emptyTextureView
+				resource: this._normalTexture ? this._normalTexture._GPUTextureView : this.redGPUContext.state.emptyTextureView
 			},
 			{
 				binding: 7,
-				resource: this._specularTexture ? this._specularTexture.GPUTextureView : this.redGPUContext.state.emptyTextureView
+				resource: this._specularTexture ? this._specularTexture._GPUTextureView : this.redGPUContext.state.emptyTextureView
 			},
 			{
 				binding: 8,
-				resource: this._emissiveTexture ? this._emissiveTexture.GPUTextureView : this.redGPUContext.state.emptyTextureView
+				resource: this._emissiveTexture ? this._emissiveTexture._GPUTextureView : this.redGPUContext.state.emptyTextureView
 			}
 
 		];
