+++ src/material/system/RedPBRMaterial_System.js
@@ -2,7 +2,7 @@
  *   RedGPU - MIT License
  *   Copyright (c) 2019 ~ By RedCamel( webseon@gmail.com )
  *   issue : https://github.com/redcamel/RedGPU/issues
- *   Last modification time of this file - 2019.12.14 17:33:43
+ *   Last modification time of this file - 2019.12.17 9:45:10
  *
  */
 
@@ -489,7 +489,7 @@ export default class RedPBRMaterial_System extends RedMix.mix(
 
 	checkTexture(texture, textureName) {
 		if (texture) {
-			if (texture.GPUTexture) {
+			if (texture._GPUTexture) {
 				switch (textureName) {
 					case 'diffuseTexture' :
 						this._diffuseTexture = texture;
@@ -510,7 +510,7 @@ export default class RedPBRMaterial_System extends RedMix.mix(
 						this._occlusionTexture = texture;
 						break;
 				}
-				if (RedGPUContext.useDebugConsole) console.log("로딩완료or로딩에러확인 textureName", textureName, texture ? texture.GPUTexture : '');
+				if (RedGPUContext.useDebugConsole) console.log("로딩완료or로딩에러확인 textureName", textureName, texture ? texture._GPUTexture : '');
 				cancelAnimationFrame(this.#raf)
 				this.#raf = requestAnimationFrame(_=>{this.needResetBindingInfo = true})
 
@@ -542,7 +542,7 @@ export default class RedPBRMaterial_System extends RedMix.mix(
 			},
 			{
 				binding: 2,
-				resource: this._displacementTexture ? this._displacementTexture.GPUTextureView : this.redGPUContext.state.emptyTextureView
+				resource: this._displacementTexture ? this._displacementTexture._GPUTextureView : this.redGPUContext.state.emptyTextureView
 			},
 			{
 				binding: 3,
@@ -558,7 +558,7 @@ export default class RedPBRMaterial_System extends RedMix.mix(
 			},
 			{
 				binding: 5,
-				resource: this._diffuseTexture ? this._diffuseTexture.GPUTextureView : this.redGPUContext.state.emptyTextureView
+				resource: this._diffuseTexture ? this._diffuseTexture._GPUTextureView : this.redGPUContext.state.emptyTextureView
 			},
 			{
 				binding: 6,
@@ -566,7 +566,7 @@ export default class RedPBRMaterial_System extends RedMix.mix(
 			},
 			{
 				binding: 7,
-				resource: this._normalTexture ? this._normalTexture.GPUTextureView : this.redGPUContext.state.emptyTextureView
+				resource: this._normalTexture ? this._normalTexture._GPUTextureView : this.redGPUContext.state.emptyTextureView
 			},
 			{
 				binding: 8,
@@ -574,7 +574,7 @@ export default class RedPBRMaterial_System extends RedMix.mix(
 			},
 			{
 				binding: 9,
-				resource: this._roughnessTexture ? this._roughnessTexture.GPUTextureView : this.redGPUContext.state.emptyTextureView
+				resource: this._roughnessTexture ? this._roughnessTexture._GPUTextureView : this.redGPUContext.state.emptyTextureView
 			},
 			{
 				binding: 10,
@@ -582,7 +582,7 @@ export default class RedPBRMaterial_System extends RedMix.mix(
 			},
 			{
 				binding: 11,
-				resource: this._emissiveTexture ? this._emissiveTexture.GPUTextureView : this.redGPUContext.state.emptyTextureView
+				resource: this._emissiveTexture ? this._emissiveTexture._GPUTextureView : this.redGPUContext.state.emptyTextureView
 			},
 			{
 				binding: 12,
@@ -590,7 +590,7 @@ export default class RedPBRMaterial_System extends RedMix.mix(
 			},
 			{
 				binding: 13,
-				resource: this._environmentTexture ? this._environmentTexture.GPUTextureView : this.redGPUContext.state.emptyCubeTextureView
+				resource: this._environmentTexture ? this._environmentTexture._GPUTextureView : this.redGPUContext.state.emptyCubeTextureView
 			},
 			{
 				binding: 14,
@@ -598,7 +598,7 @@ export default class RedPBRMaterial_System extends RedMix.mix(
 			},
 			{
 				binding: 15,
-				resource: this._occlusionTexture ? this._occlusionTexture.GPUTextureView : this.redGPUContext.state.emptyTextureView
+				resource: this._occlusionTexture ? this._occlusionTexture._GPUTextureView : this.redGPUContext.state.emptyTextureView
 			}
