+++ src/material/system/RedSkyBoxMaterial.js
@@ -2,7 +2,7 @@
  *   RedGPU - MIT License
  *   Copyright (c) 2019 ~ By RedCamel( webseon@gmail.com )
  *   issue : https://github.com/redcamel/RedGPU/issues
- *   Last modification time of this file - 2019.12.14 16:4:46
+ *   Last modification time of this file - 2019.12.17 9:45:10
  *
  */
 
@@ -73,13 +73,13 @@ export default class RedSkyBoxMaterial extends RedMix.mix(
 
 	checkTexture(texture, textureName) {
 		if (texture) {
-			if (texture.GPUTexture) {
+			if (texture._GPUTexture) {
 				switch (textureName) {
 					case 'skyBoxTexture' :
 						this._skyBoxTexture = texture;
 						break
 				}
-				if (RedGPUContext.useDebugConsole) console.log("로딩완료or로딩에러확인 textureName", textureName, texture ? texture.GPUTexture : '');
+				if (RedGPUContext.useDebugConsole) console.log("로딩완료or로딩에러확인 textureName", textureName, texture ? texture._GPUTexture : '');
 				this.needResetBindingInfo = true
 			} else {
 				texture.addUpdateTarget(this, textureName)
@@ -96,7 +96,7 @@ export default class RedSkyBoxMaterial extends RedMix.mix(
 			{binding: 0, resource: this.sampler.GPUSampler},
 			{
 				binding: 1,
-				resource: this._skyBoxTexture ? this._skyBoxTexture.GPUTextureView : this.redGPUContext.state.emptyCubeTextureView
+				resource: this._skyBoxTexture ? this._skyBoxTexture._GPUTextureView : this.redGPUContext.state.emptyCubeTextureView
 			}
 		];
 		this._afterResetBindingInfo();
