+++ src/resources/RedBitmapTexture.js
@@ -2,58 +2,19 @@
  *   RedGPU - MIT License
  *   Copyright (c) 2019 ~ By RedCamel( webseon@gmail.com )
  *   issue : https://github.com/redcamel/RedGPU/issues
- *   Last modification time of this file - 2019.12.16 20:34:19
+ *   Last modification time of this file - 2019.12.17 9:45:10
  *
  */
 "use strict";
 import RedSampler from "./RedSampler.js";
 import RedUUID from "../base/RedUUID.js";
 import RedImageLoader from "./system/RedImageLoader.js";
-import RedGPUContext from "../RedGPUContext.js";
+import RedCopyBufferToTexture from './system/RedCopyBufferToTexture.js'
+import RedBaseTexture from "../base/RedBaseTexture.js";
 
 
 let defaultSampler;
 const MIPMAP_TABLE = new Map();
-const updateTexture = function (commandEncoder, device, imageDatas, gpuTexture, updateTarget, face = -1) {
-	let promise = new Promise(((resolve, reject) => {
-
-		imageDatas.forEach((info, mip) => {
-			if (!updateTarget.useMipmap && mip) return
-			let data = new Uint8ClampedArray(info.data)
-			let width = info.width
-			let height = info.height
-			let rowPitch = info.rowPitch;
-			const textureDataBuffer = device.createBuffer({
-				size: data.byteLength + data.byteLength % 4,
-				usage: globalThis.GPUBufferUsage.COPY_DST | globalThis.GPUBufferUsage.COPY_SRC,
-			});
-			// console.log(imageData)
-			textureDataBuffer.setSubData(0, data);
-			const bufferView = {
-				buffer: textureDataBuffer,
-				rowPitch: rowPitch,
-				imageHeight: height,
-			};
-			const textureView = {
-				texture: gpuTexture,
-				mipLevel: mip,
-				arrayLayer: Math.max(face, 0),
-			};
-
-			const textureExtent = {
-				width: width,
-				height: height,
-				depth: 1
-			};
-			commandEncoder.copyBufferToTexture(bufferView, textureView, textureExtent);
-			if (RedGPUContext.useDebugConsole) console.log('mip', mip, 'width', width, 'height', height)
-		})
-
-		resolve()
-	}))
-
-	return promise
-};
 let makeMipmap = function (redGPUContext, imageDatas, targetTexture) {
 	console.log('imageDatas', imageDatas)
 	let tW = imageDatas[0].width;
@@ -76,7 +37,7 @@ let makeMipmap = function (redGPUContext, imageDatas, targetTexture) {
 	const gpuTexture = redGPUContext.device.createTexture(textureDescriptor);
 	MIPMAP_TABLE.set(targetTexture.mapKey, gpuTexture)
 	// console.log(tW, tH)
-	updateTexture(commandEncoder, redGPUContext.device, imageDatas, gpuTexture, targetTexture).then(
+	RedCopyBufferToTexture(commandEncoder, redGPUContext.device, imageDatas, gpuTexture, targetTexture).then(
 		_ => {
 			targetTexture.resolve(gpuTexture)
 			if (targetTexture.onload) targetTexture.onload()
@@ -84,10 +45,8 @@ let makeMipmap = function (redGPUContext, imageDatas, targetTexture) {
 		}
 	)
 }
-export default class RedBitmapTexture extends RedUUID {
-	#updateList = [];
-	#GPUTexture;
-	#GPUTextureView;
+export default class RedBitmapTexture extends RedBaseTexture {
+
 	//TODO 옵션값이 바뀌면 어떻게 할건지 결정해야함
 	constructor(redGPUContext, src, sampler, useMipmap = true, onload, onerror) {
 		super()
@@ -119,28 +78,7 @@ export default class RedBitmapTexture extends RedUUID {
 		}
 	}
 
-	get GPUTexture() {
-		return this.#GPUTexture
-	}
 
-	get GPUTextureView() {
-		return this.#GPUTextureView
-	}
-
-	resolve(texture) {
-		this.#GPUTexture = texture;
-		this.#GPUTextureView = texture ? texture.createView() : null;
-		this.#updateList.forEach(data => {
-			data[0][data[1]] = this
-		});
-		this.#updateList.length = 0
-	}
-
-	addUpdateTarget(target, key) {
-		this.#updateList.push([
-			target, key
-		])
-	}
 
 
 }
